etl/gdrive_to_s3/requirements.txt
google-api-python-client
google-auth
boto3
awslambdaric
